This one runs.

+++ src/main/java/com/example/Biografen/Objects/ShiftRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ShiftRepository extends JpaRepository<Shift, Integer> {
+public interface ShiftRepository extends JpaRepository<Shift, Long> {
     List<Shift> findByNameStartsWithIgnoreCase (String name);
 }
